+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 - Updated the "Unknown label" error message to match other error messages
   (#1548)
 - Permit type holes in function arguments and return annotations (#1519)
+- Report unused module imports (#1553)
 
 ## v0.20.1 - 2022-02-24